Move recording parsing to its own file; warn if recording version low, error if high
